Add doc + functions in doc order
